Move mmr_GetNewGiIndex out of replace-gi-table mod and into mmr.c along with bottle replacement logic. Update gi_index usages to be u16 instead of u32.

+++ MMR.Randomizer/Extensions/ItemExtensions.cs
@@ -10,7 +10,7 @@ namespace MMR.Randomizer.Extensions
 {
     public static class ItemExtensions
     {
-        public static int? GetItemIndex(this Item item)
+        public static ushort? GetItemIndex(this Item item)
         {
             return item.GetAttribute<GetItemIndexAttribute>()?.Index;
         }
